michelinDictator/requirements.txt
apturl==0.5.2
asgiref==3.5.2
astor==0.8.1
attrs==22.1.0
autopep8==2.0.0
bandit==1.7.2
bcrypt==3.2.0
blinker==1.4
blis==0.7.9
Brlapi==0.8.3
catalogue==2.0.8
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
confection==0.0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
darglint==1.8.1
DAWG-Python==0.7.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-email-users==1.1.0
django-filter==22.1
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
duplicity==0.8.21
eradicate==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0
fasteners==0.14.1
filelock==3.8.0
flake8==3.9.0
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==21.11.29
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
isort==5.10.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.2
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.2.1
murmurhash==1.0.9
mypy==0.991
mypy-extensions==0.4.3
netifaces==0.11.0
nltk==3.7
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
paramiko==2.9.3
pathy==0.8.1
pbr==5.11.0
pep8-naming==0.11.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.4
pluggy==1.0.0
preshed==3.0.8
protobuf==3.12.4
psycopg2==2.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycodestyle==2.7.0
pycups==2.0.1
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-math==0.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
restructuredtext-lint==1.4.0
SecretStorage==3.3.1
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.5
stevedore==4.1.0
sympy==1.11.1
systemd-python==234
texttable==1.6.4
thinc==8.1.5
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing==3.7.4.3
typing_extensions==4.4.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.7
wadllib==1.3.6
wasabi==0.10.1
websocket-client==1.4.2
wemake-python-styleguide==0.15.3
xdg==5
xkit==0.0.0
zipp==1.0.0
